build: compile clean under clang, not just MSVC
Worked against the pre-move layout: capture_wasapi.cpp and passthrough_wasapi.cpp
are still under src/lib/src/{capture,sinks}/ here. The sibling task that moves
them into a per-platform tree will need to carry these two files across.

No preprocessor branching added anywhere, and cmake/CompilerWarnings.cmake is
untouched - the strict set still applies in full to every file under src/.

  __uuidof (15 sites)   An MSVC extension; -Wpedantic rejects it. The obvious
                        substitution - the SDK's CLSID_/IID_ constants - does
                        not link: mmdeviceapi.h and audioclient.h only declare
                        them (EXTERN_C const, not DEFINE_GUID, so <initguid.h>
                        cannot help either) and no shipped import library
                        defines them. Both files already spell GUIDs out as
                        constexpr literals for exactly this class of reason
                        (kPkeyDeviceFriendlyName, kSubtypeIec61937DolbyDigital),
                        so these follow suit. IID_PPV_ARGS expands to __uuidof
                        too, and KSDATAFORMAT_SUBTYPE_PCM / _IEEE_FLOAT expand
                        to it via mmreg.h's DEFINE_GUIDNAMED. All seven values
                        were checked against the SDK's own __uuidof at runtime
                        before being written down.

  -Wsign-conversion     int used as a container index, in tables.hpp,
                        bitwriter.hpp, silent_frame.hpp and bitalloc.cpp.
                        Widened the loop variable where the index was the only
                        reason it was int; cast at the use site otherwise,
                        matching bitalloc.cpp's existing style.

  -Wdouble-promotion    A const float silently widened inside a double
                        expression, in atmos.cpp and spatial.cpp. Made the
                        widening explicit. The arithmetic is unchanged: it was
                        already being done in double.

  missing initializer   ScannedStream::access_units and joc::FrameParameters::
                        matrix were the only members of their structs without a
                        default. Not a latent bug - every site either resizes
                        immediately or relies on the vector starting empty
                        (scan_eac3 tests access_units.empty()), which is what
                        the omitted designator already gave them. Added {} so
                        the aggregate is uniformly defaulted.

Two things surfaced behind these once the errors stopped hiding them.

`play` had been dropped. acd2c78 rebuilt dispatch as a table and left run_play
out of it, so the command stopped existing - no usage line, no dispatch - while
the function stayed. Nothing caught it: clang skips -Wunused-function in a
translation unit that already has errors, and the errors above were in headers
main.cpp includes. Restored as a 17th row. Its device_index default is -1, not
0, because run_play reads -1 as "the default endpoint"; and min_args now bounds-
checks the argument, which the old `if (command == "play")` branch did not - it
read args[1] unconditionally.

qmlcachegen's output is Qt's code, regenerated every build and not editable: its
value registers are int while QJSList::indexOf returns qsizetype, which
-Wconversion rejects. Qt's headers are already exempt from ac3::warnings via
SYSTEM includes; src/gui/CMakeLists.txt now extends that to what Qt generates,
and to nothing else. Clang frontends only - cl.exe is already quiet there and a
second warning-level flag would just trade the clean build for a screen of D9025.

The tests needed the same four fixes; the 29 errors in the report were the
ac3cli target alone, which does not build them.

Verified both compilers from a clean configure. clang 21.1.0: builds with no
errors and no warnings, 175/175 pass. MSVC 14.51: builds with no errors and no
warnings, 175/175 pass.

Behaviour is unchanged, and checked rather than assumed: ac3cli built from
acd2c78 and from this commit produce byte-identical output over 18 workloads -
AC-3 5.1 and 7.1 sine, orbit panning, Atmos objects, E-AC-3 at 5.1/7.1.4 with
all coding tools, silence, decode, re-encode, spdif and mkv wrapping. levels and
loudness agree on every figure they print, differing only in the input path they
echo back.

# Conflicts:
#	src/cli/main.cpp
#	src/gui/CMakeLists.txt

+++ src/lib/include/ac3/core/bitwriter.hpp
@@ -36,7 +36,9 @@ public:
         }
     }
 
-    [[nodiscard]] std::size_t bit_count() const { return bytes_.size() * 8 + pending_; }
+    [[nodiscard]] std::size_t bit_count() const {
+        return bytes_.size() * 8 + static_cast<std::size_t>(pending_);
+    }
 
     // Overwrite a 16-bit big-endian field at a byte offset in already-emitted
     // output. Used to patch the two CRC words after the frame body is packed
